Remove reflection from Environment.WorkingSet prop
